branch broke had to fix
it seems like it overwrite just about everything that I had done on the GUI side, so I went ahead and replaced all of the relavant code. (This was just under the Gameserver, littleServer, and oneForAllScript, so all GUI related). I then updated the back end by adding a "is_hunting_hare" variable and updating the funtion so it SHOULD consult it. give me a second to test it and see if it works, if you see nothing else about it , assume it works.