low level json aqquired
created and deployed the low level json. Should it save to somewhere other than my desktop? definititely! why haven't I dont that yet? no clue. Also worked on the high level json, there are some tweaks to be made and some bugs (specifically the zero movement bug) but we are getting close to done! also revisit the high level dict, make it not so hard coded. make it softer. maybe.